Added if protection to make the `destroy' safe during GC.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@70270 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/library/docking/implementation/toolbar/sd_tool_bar_zone.e
@@ -332,7 +332,9 @@ feature -- Command
 			-- Destroy
 		do
 			internal_shared.widgets.prune_tool_bar_zone (Current)
-			tool_bar.destroy
+			if tool_bar /= Void then
+				tool_bar.destroy
+			end
 			Precursor {IDENTIFIED}
 		end
